To AS
---

Added a "global" option to msr2data

First, I hope, this major "surgery" does not cause any regressions in the standard features.
If so, please revert to an older version and file a bug report.

This new option is not yet documented and should be considered as experimental.

Also the newly added code still needs some cleanup - many things have simply been written as a proof of concept.
This will be done during further testing and fine-tuning stages.