Integrated building and installing the Qt editor into the musrfit installation. At the moment this is done in the simple way that qmake is called from the configure script.
Since there is not really a straightforward way to look for Qt installations at certain paths,
the automatic determination of the available Qt version is only done through pkg-config.
In case Qt is found at non-standard installation paths, one can either use the configure options
"--with-qt3" or "--with-qt4" to specify the Qt directory or alternatively set the variable
PKG_CONFIG_PATH to some value like
/usr/lib/pkgconfig:/usr/local/lib/pkgconfig:$ROOTSYS/lib/pkgconfig:/opt/qtsdk-2010.02/qt/lib/pkgconfig

During the installation only one editor---either musredit or musrgui---is built and installed.
musredit/Qt4 is generally preferred over musrgui/Qt3.
The only way to install musrgui when also a sufficent Qt4 installation is present is to specify solely
the "--with-qt3" option on the configure level. If additionally the "--with-qt4" option is given, only
musredit will be installed.

Both editors still can be installed as previously---this step is merely to make the installation more
convenient for less-experienced users (hopefully).