AddT0 offset bug
Up to now there has been an addT0 offset bug which resulted to the following situation.
Since the offset was wrong by '1', for the first addt0 the t0 of the file was used
instead of the one given in the msr-file. If these t0's are substantially different, the
adding of the runs was quite off, resulting in strong distortions at very early times.
Pant Amba Datt pointed towards this bug which is fixed now.